Add shared library to monorepo
