+++ package.json
@@ -23,8 +23,8 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "coverage": "vitest run --coverage",
-    "check:types": "tsc --build",
-    "checks": "pnpm clean && pnpm run '/^check:.*/'",
+    "check:types": "tsc --build --force --verbose",
+    "checks": "pnpm run '/^check:.*/'",
     "start": "pnpm --filter \"graph-explorer-proxy-server\" run start",
     "clean": "pnpm --stream -r run clean",
     "build": "pnpm --stream -r run build",
