+++ packages/graph-explorer-proxy-server/package.json
@@ -14,6 +14,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@aws-sdk/credential-providers": "^3.621.0",
+    "@graph-explorer/shared": "workspace:*",
     "aws4": "^1.13.0",
     "body-parser": "^1.20.2",
     "compression": "^1.7.4",
