+++ packages/graph-explorer/package.json
@@ -20,6 +20,7 @@
     "@emotion/react": "^11.11.4",
     "@emotion/serialize": "^1.1.4",
     "@emotion/utils": "^1.2.1",
+    "@graph-explorer/shared": "workspace:*",
     "@mantine/core": "^7.9.0",
     "@mantine/dates": "^7.9.0",
     "@mantine/emotion": "^7.9.0",
