+++ packages/graph-explorer/src/connector/sparql/sparqlExplorer.ts
@@ -20,7 +20,7 @@ import {
   SPARQLKeywordSearchRequest,
   SPARQLNeighborsRequest,
 } from "./types";
-import { ConnectionConfig } from "@/core";
+import { ConnectionConfig } from "@shared/types";
 import { v4 } from "uuid";
 import { env, logger } from "@/utils";
