+++ packages/graph-explorer/src/hooks/useEntities.test.tsx
@@ -4,8 +4,6 @@ import { Vertex } from "@/types/entities";
 import {
   createRandomEdge,
   createRandomEntities,
-  createRandomInteger,
-  createRandomName,
   createRandomSchema,
   createRandomVertex,
 } from "@/utils/testing";
@@ -16,6 +14,7 @@ import { Entities } from "@/core/StateProvider/entitiesSelector";
 import { renderHookWithRecoilRoot } from "@/utils/testing";
 import { waitForValueToChange } from "@/utils/testing/waitForValueToChange";
 import { vi } from "vitest";
+import { createRandomInteger, createRandomName } from "@shared/utils/testing";
 
 describe("useEntities", () => {
   beforeEach(() => {
