+++ packages/graph-explorer/src/modules/CreateConnection/CreateConnection.tsx
@@ -6,9 +6,9 @@ import { InfoTooltip } from "@/components";
 import Button from "@/components/Button";
 import Input from "@/components/Input";
 import Select from "@/components/Select";
+import { ConnectionConfig } from "@shared/types";
 import {
   ConfigurationContextProps,
-  ConnectionConfig,
   RawConfiguration,
   useWithTheme,
 } from "@/core";
