+++ packages/graph-explorer/src/utils/testing/randomData.ts
@@ -1,4 +1,3 @@
-import { v4 } from "uuid";
 import {
   AttributeConfig,
   EdgeTypeConfig,
@@ -8,6 +7,15 @@ import {
 } from "@/core";
 import { Edge, Vertex } from "@/types/entities";
 import { Entities } from "@/core/StateProvider/entitiesSelector";
+import {
+  createArray,
+  createRandomBoolean,
+  createRandomInteger,
+  createRandomName,
+  createRandomUrlString,
+  createRecord,
+  randomlyUndefined,
+} from "@shared/utils/testing";
 
 /*
 
@@ -22,102 +30,6 @@ The randomness of all the other values ensures that the logic under test is not
 affected by those values, regardless of what they are.
 
 */
-
-/**
- * Creates a random string with a prefix, if provided.
- * @param prefix The prefix to prepend to the string.
- * @returns A random string that will resemble "prefix-8d49f0".
- */
-export function createRandomName(prefix: string = ""): string {
-  return `${prefix}${prefix.length > 0 ? "-" : ""}${v4().substring(0, 6)}`;
-}
-
-/**
- * Creates a random boolean.
- * @returns A random boolean value.
- */
-export function createRandomBoolean(): boolean {
-  return Math.random() < 0.5;
-}
-
-/**
- * Creates a random integer.
- * @param max The maximum value the random integer can have. Defaults to 100,000.
- * @returns A random integer value from 0 to the max.
- */
-export function createRandomInteger(max: number = 100000): number {
-  return Math.floor(Math.random() * max);
-}
-
-/**
- * Randomly creates a hex value for an RGB color.
- * @returns The hex string of the random color.
- */
-export function createRandomColor(): string {
-  const letters = "0123456789ABCDEF".split("");
-  let color = "#";
-  for (let i = 0; i < 6; i++) {
-    color += letters[Math.round(Math.random() * 15)];
-  }
-  return color;
-}
-
-/**
- * Randomly creates a URL string.
- * @returns The URL string.
- */
-export function createRandomUrlString(): string {
-  const scheme = pickRandomElement(["http", "https"]);
-  const host = createRandomName("host");
-  const port = pickRandomElement(["", `:${createRandomInteger(30000)}`]);
-  const path = pickRandomElement(["", `/${createRandomName("path")}`]);
-  return `${scheme}://${host}${port}${path}`;
-}
-
-/**
- * Creates a random date.
- * @param start The lower bound of the random date.
- * @param end The upper bound of the random date.
- * @returns A random `Date` value bound by the given `start` and `end` values.
- */
-export function createRandomDate(start?: Date, end?: Date): Date {
-  const startTime = start ? start.getTime() : new Date(1970, 0, 1).getTime();
-  const endTime = end ? end.getTime() : new Date().getTime();
-  const randomTime = startTime + Math.random() * (endTime - startTime);
-  return new Date(randomTime);
-}
-
-/**
- * Randomly returns the provided value or undefined.
- * @returns Either the value or undefined.
- */
-export function randomlyUndefined<T>(value: T): T | undefined {
-  return createRandomBoolean() ? value : undefined;
-}
-
-/**
- * Creates an array containing values generated from the factory function with the given length.
- * @param length The number of items to generate.
- * @param factory A function to generate the desired value.
- * @returns An array with items generated using the factory function.
- */
-export function createArray<T>(length: number, factory: () => T): T[] {
-  return Array.from({ length }, factory);
-}
-
-export function createRecord<TValue>(
-  length: number,
-  factory: () => { key: string; value: TValue }
-): Record<string, TValue> {
-  const result: Record<string, TValue> = {};
-
-  for (let i = 0; i < length; i++) {
-    const newEntry = factory();
-    result[newEntry.key] = newEntry.value;
-  }
-
-  return result;
-}
 
 /**
  * Creates a random AttributeConfig object.
