+++ packages/shared/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "extends": ["../../tsconfig.base.json"],
+  "compilerOptions": {
+    "composite": true,
+    "lib": ["es2023"],
+    "module": "node16",
+    "target": "es2022",
+    "types": ["vitest/globals"],
+
+    /* Path aliasing */
+    "outDir": "./dist",
+    "baseUrl": "./src",
+    "rootDir": "./src",
+
+    /* Bundler mode */
+    "moduleResolution": "NodeNext",
+    "emitDeclarationOnly": true
+  },
+  "include": ["src"]
+}
