+++ packages/shared/vitest.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    globals: true,
+    environment: "node",
+    coverage: {
+      reportsDirectory: "coverage",
+      provider: "v8",
+      reporter: ["lcov", "text", "json", "clover"],
+    },
+  },
+});
